First React Native demo app - Is a Todo List App, just add items to a list and display all of those items
Components Use:
- Flat List (React-Native)
- useState (React)
- types (JS/TS)
- Interfaces (TS)
- Pressable
- Modal

New:
- Added Modal with a new button at the start to show the Todo List. With this modal encountered differences about ios and android styling and had to make their own stylesheets. So created a `DeviceStyleManger` to pass ass parameters both device style and handle the differences in styling depending on device.
Also, created a string resource file to access all the string values. Created this because the text "Add Todo List" the 'todo' part trigger the TODO keyword on the compiler and would make it as an actual TODO notation. This prevents the notation trigger

Currently, has 2 functions :
- GetDeviceViewStyling()
- GetDeviceTextStyling()